# Corgiverse-AHA-Specification-v1.0.0/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AHA Local API",
    "version": "1.0.0",
    "description": "Build contract; no running server is supplied. All references bundled."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/v1"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/session": {
      "post": {
        "operationId": "pairSession",
        "summary": "pairSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          }
        },
        "security": []
      },
      "delete": {
        "operationId": "logout",
        "summary": "logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/cases": {
      "get": {
        "operationId": "listCases",
        "summary": "listCases",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 20000
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasePage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createCase",
        "summary": "createCase",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCase"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Case": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "schema_version": {
            "const": "1.0.0"
          },
          "case_revision": {
            "type": "integer",
            "minimum": 0
          },
          "synthetic": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "timezone",
          "schema_version",
          "case_revision",
          "synthetic",
          "created_at"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "field_errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200
                },
                "code": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200
                },
                "message": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 20000
                }
              },
              "required": [
                "path",
                "code",
                "message"
              ],
              "additionalProperties": false
            },
            "minItems": 0,
            "maxItems": 10000
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "field_errors",
          "retryable"
        ],
        "additionalProperties": false
      },
      "CreateCase": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "synthetic": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "timezone",
          "synthetic"
        ],
        "additionalProperties": false
      },
...
      "CasePage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Case"
            },
            "minItems": 0,
            "maxItems": 200
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "additionalProperties": false
      },
...
      "SessionRequest": {
        "type": "object",
        "properties": {
          "pairing_secret": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "operator_label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "pairing_secret",
          "operator_label"
        ],
        "additionalProperties": false
      },
      "Session": {
        "type": "object",
        "properties": {
          "actor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "csrf_token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "actor",
          "csrf_token",
          "expires_at"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aha_session"
      },
      "gatewayToken": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}